Improved Dialogue and Dialogue healing
All within TestSceneCollision:
Added having to press a button to initiate dialogue (new button is enter, not space).
Changed inputs to use buttons within Unity, so will be easier to change in the future.
Added a new test NPC that heals the player to full (meant to test NPC's affecting the player, NPC is located in the top right. Also has placeholder sprite).
Changed dodge to now only have Dave move in the direction he was already moving, making it more of a roll. Will be hard to make feel right without animations.